+++ src/packages/piral-core/src/state/actions/index.ts
@@ -1,3 +1,4 @@
 export * from './app';
 export * from './components';
 export * from './data';
+export * from './define';
